requirements.txt
# Core dependencies used in src/
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
scikit-learn>=1.2.0

# Required by pandas for datetime handling
python-dateutil>=2.8.0
pytz>=2023.3

# Forecasting pipeline (5 models)
xgboost>=2.0.0
statsmodels>=0.14.0
torch>=2.0.0

# News pipeline (crawl, clean, sentiment, align)
beautifulsoup4>=4.12.0
PyYAML>=6.0

# Web app (FastAPI + static frontend)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
